agent/open_deep_research/requirements.txt
langgraph>=0.2.55
langchain-community>=0.3.9
langchain-openai>=0.3.7
langchain-anthropic>=0.3.9
openai>=1.61.0
tavily-python>=0.5.0
langchain-groq>=0.2.4
arxiv>=2.1.3
pymupdf>=1.25.3
xmltodict>=0.14.2
linkup-sdk>=0.2.3
duckduckgo-search>=3.0.0
exa-py>=1.8.8
requests>=2.32.3
beautifulsoup4==4.13.3
langchain-deepseek>=0.1.2
python-dotenv>=1.0.1
langgraph_supervisor
langchain_tavily
pytest
httpx>=0.24.0
markdownify>=0.11.6
azure-identity>=1.21.0
azure-search>=1.0.0b2
azure-search-documents>=11.5.2